add process bar

+++ helpers.py
@@ -1,12 +1,30 @@
 import os, shutil
 import requests
 import sys
-
 from zipfile import ZipFile
+
+import bcolor
+
+def process_bar(total_length, dl_process):
+    done = int(50 * dl_process / total_length)
+    
+    
+    process = "\r{0}Downloading...{1} [{2}{3}>{4}{5}] {6}%".format(
+        bcolor.Bcolor.OKGREEN,
+        bcolor.Bcolor.ENDC,
+        bcolor.Bcolor.CRED2,
+        '=' * done,
+        ' ' * (50-done),
+        bcolor.Bcolor.ENDC,
+        done * 2
+    )
+    sys.stdout.write(process)    
+    sys.stdout.flush()
+    
 
 def download(url: str, save_path: str = "data", file_name: str = "data.zip"):
     path_file = os.path.join(save_path, file_name)
-    dl = 0
+    dl_process = 0
     
     try:
         with open(path_file, "wb") as f:
@@ -18,11 +36,10 @@ def download(url: str, save_path: str = "data", file_name: str = "data.zip"):
                 return True
             
             for data in res.iter_content(chunk_size=4096):
-                dl += len(data)
                 f.write(data)
-                done = int(50 * dl / total_length)
-                sys.stdout.write("\rDownloading [%s>%s]" % ('=' * done, ' ' * (50-done)) )    
-                sys.stdout.flush()
+                
+                dl_process += len(data)
+                process_bar(total_length, dl_process)
 
             f.close()
 
@@ -43,14 +60,18 @@ def unzip(zip_file: str, save_folder: str = "./data", new_file: str = "data.json
             zip_ref.extractall(save_folder)
 
         os.rename(old_file, new_file)
-        print(old_file)
-        os.remove(zip_file)
+
+        # remove file zip
+        if os.path.exists(zip_file):
+            print("Rm file " + zip_file)
+            os.remove(zip_file)
     except:
         return None
     
 def rm_all_file_in_folder(folder: str):
     for filename in os.listdir(folder):
         file_path = os.path.join(folder, filename)
+        
         try:
             if os.path.isfile(file_path) or os.path.islink(file_path):
                 os.unlink(file_path)
